fork: Add option to skip the clone
The ability to skip the clone is needed, for the work flow where you
already have a clone of the origin version and you are deciding to
push a new branch or change to your local repository.  It is
especially useful if you have shallow cloned or cloned only a single
branch out of a large repo.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>