lab: Add --comments flag to 'mr show'
The 'issue show' command has a --comments flag to show all the comments
on an issue.  The same command is useful for the merge request command
'mr show'.

Add a --comments flag to 'mr show' that displays all the comments made
on a merge request.

Minor fix: Also fix a typo with the mr no-markdown flag.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>