mr show: Add --since option
The show command shows all comments in the default font which makes
finding new comments difficult on merge requests with many comments.
It is useful to have a mechanism to highlight, in bold text, new
comments.

Add a --since option which allows users to specify a date to highlight
comments added after that date.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>