show: Mark updated comments
Currently updated comments look like other comments.  It is useful to
note that a comment has been updated.

Update 'issue show' and 'mr show' to show the updated comment date and
output a message indicated that the comment was updated.

[Note: gitlab sets the note's UpdatedAt equal to CreatedAt at the time
of creation so it is okay to make the UpdateAt to CreatedAt comparison
and only use UpdatedAt in the output.]

Signed-off-by: Prarit Bhargava <prarit@redhat.com>