mr create: Add mr create --file option
Add the ability to create a merge request with the contents of a file.
This is especially useful when creating a merge request with testing
contents and/or boot logs.

Add mr create --file option which takes the location of the file containing
the text as an argument.  The first line of the file will be interpreted as
the title of the merge request, and the remaining text will be the body for
the description of the merge request.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>