mr show: Add automatic comment highlights
One of the features that is lost with an email based git workflow (ie the
kernel) is the ability to keep track of new comments. In a mailer this is
easy because the new comments remain unread but in a text-based console
it's not so easy to do.

While the --since option allows users to highlight new comments from the
comand line, it is a bad user experience as users have to remember the
date they last viewed an MR.

A better approach is to keep per-repo metadata in .git/labmetadata.hcl
that keeps record of the last time that a merge request was viewed by the
user.  This way only comments that were added since the last user access
are highlighted without requiring user intervention.

The format of the metadata for issues is:

"mr<number>" "time stamp"

For example,

	"mr9" = "2020-08-26T18:43:33.739792801Z"

Add automatic comment highlights for 'mr show'.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>