[PATCH] Add `--patch` option to `mr show`
Fixes #277 which requests functionality that provides the patches
for an MR in 'git show' format.  This is useful for users that want
to quickly look at a patchset without checking out a branch.

The original patch provided by Eric required users to provide the
target remote, even if the target remote was "origin".  The additional
code provided by Prarit uses the fetch remote name in 'git remote -v'
by default unless the user specifies a specific remote.

Original patch from Eric, with modifications to automatically find the
target remote, and the test from Prarit.

Add '--patch' option to 'mr show' that displays the patches for a merge
request in 'git show' format.

Co-Developed-by: Eric Engestrom <eric.engestrom@intel.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>