config: Rework the config code
The existing config code uses a lot of workarounds for viper's issues with
the HCL format.  With the move TOML format for the config files, a lot of
this code can be eliminated by using standard viper calls.

Rework the config code to remove workarounds and reorganize the config
function.

Suggested-by: Zaq? Wiedmann <zaquestion@gmail.com>
Signed-off-by: Prarit Bhargava <prarit@redaht.com>