config: Fix and clean up certificate code
As noted in issue #152 ("Accept self signed certificates"), the
certificate code is no longer working.  This is because the certificate
code was executed before the configuration was loaded.

The code is need of a massive clean up, and with the transition of the
config files to TOML format, the code to get the certificate can easily be
incorporated into the main configuration code.

Fix and clean up the certifcate code.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>