config: Read token from other sources
Currently the gitlab personal access token is stored in cleartext in
the config file.  Allowing for users to specify a command to fetch the
token from a password manager or other encryption utility provides
additional security.

Add a "load_token" config entry to allow users to specify another utility
to obtain the token.

Suggested-by: Thomas Furfaro <tom@furfmon.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>