(gitlab-ci) add in ssh config and override CI config mode
When running lab in GitLab CI, lab will attempt to autoconfigure to use
the CI credentials. While this is a useful feature for consumers of lab,
it's working against us here. I've used the lab env vars which currently
take precedence over the CI config, but when #414 comes in, I believe
it'll break again, so we'll need to revisit then.

The ssh config make it so lab can clone repos over ssh, as that key is
attached to the `lab-testing` user