show: Add "comments" config option.
I'd like to always display the full comments when viewing an MR or Issue.
Add a "comments" config variable that sets the "--comments" flag on
'mr show' and 'issue show'.

Add a "comments" local config option for the show commmands.

Additional fix: Fix a leftover HCL to TOML change in the README

Signed-off-by: Prarit Bhargava <prarit@redhat.com>