issue,mr: remove markdown flag from parent command
The flag "--no-markdown","-M" was being shown in both 'lab mr' and 'lab issue'
command as a possible flag, but it had no effect whatsoever in any commands
other than in the 'show' subcommands, which had their own implementation of
this very same flag. With that, remove it from the the parent commands to avoid
confusion on its usage.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>