lab: Add --force-linebreak option
Markdown formatting will mangle some types of text by default.

For example, if the text entered as the Description contains

This is sentence one and should not be
On the same line as sentence two

then markdown will merge these lines together as

This is sentence one and should not be On the same line as sentence two.

To fix this add a --force-linebreak option to replace line breaks, "\n",
with "  \n" [1].

Signed-off-by: Prarit Bhargava <prarit@redhat.com>

[1] https://gist.github.com/shaunlebron/746476e6e7a4d698b373.