fork: fix clonning for projects within subgroups
Projects placed within a subgroup in gitlab was not being cloned while using
'fork' cmd. Even though the project was being properly forked in gitlab, the
issue was that only the first path part (after namespace) was being used as the
project name to be clone from user's namespace, e.g.  'user/group/project' was
being handled by lab as a project with name 'group' instead of 'project'.  This
patch uses the last path part always, regardless the project being placed
within a group or not.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>