root_test: Add tests for new config and flag layout
There isn't really a good way to test the config and flag override
code.  I've opted to call 'mr show' with different flag and config
settings to test out the code.  The tests include testing for
.git/lab/lab.toml overrides of ~/.config/lab/lab.toml and tests
for the override by the flag of the value in the config files.

Add tests for the new config and flag layout.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>