Add commit discussion functionality
The GitLab UI provides the ability to show comments that were made on
patches, and this functionality is missing from lab.

Enhance 'lab mr show' to display the diffs and add a --color-diff option
to show the diffs in color (red for removals, green for additions).

Note that due to the design of the GitLab API, not all diffs are viewable.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>