root: add fallthrough option for setting forkedFromRemote
In case 'forkedFromRemote' isn't set by neither of the options:
1) 'upstream' remote is present
2) 'origin' remote is present
3) remote being tracked by 'master' (branch may not exist)

Then lab now has a last option:
4) get the first remote placed in .git/config file, which, usually, is the one
   set when the clone was performed.

The 'go-gitconfig' lib doesn't have an option for retrieving this info; the
functions git.go:Remotes() and git.go:GetLocalRemotes() return the remote
values from the 'git remote' command, which orders the output in alphabetical
order.  Because of that, this patch adds the git.go:GetLocalRemotesFromFile()
function, which get the remotes from .git/config file in placement order.  And,
finally, the option (4) is then added to cmd/root.go:Execute().

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>