root: add default values to test setup
Some tests depend on the default values of 'forkedFromRemote and 'forkRemote'
variables; these tests run internally calling the command Run() code, instead
of the launching the lab executable.  With that, the code that correctly sets
these vars is bypassed and thus they are left empty.  These test cases rely
only on the "origin" value, so setting it on the test setup should be fine.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>