root: skip test instead of fail if running locally
The Test_config_gitConfig_* tests depends on the test being running on the
Travis CI run home directory.  It doesn't point to the "current system home
dir" to avoid deleting user's config file if the test is run locally.  Because
of that, all tests fails due to "file or directory not found".  With this
patch, instead of failing, the tests are skipped if the file isn't found.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>