lab: Add MR edit functionality
It is possible to edit an issue in lab.  It would be good to have
the ability to edit an MR as well, and have the options match
the existing issue edit functionality.

The command 'lab mr edit <mr_id>' allows a user to modify the existing
Merge Request description for the specified MR ID.

Add functionality to edit a Merge Request.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>