mr create: Update 'mr create' comments
The 'lab mr create' command provides changelog output as part of the
default commit message in which each commit, author, and age of commit
is displayed.  The output does not contain a diffstat and some
users find diffstat useful for both commit verification and for code
review.

Remove the age of the commits from the output and provide the diffstat
output in the create output.  Like the rest of the changelog, the new
diffstat output is commented so it is not included in the merge request
description by default.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>