mr create: Add --cover-letter option
The 'git-format-patch' command has a nice --cover-letter option that
displays a changelog and diffstat in a patch series.  This information
can be useful for reviewers viewing MR requests so add an option to
include it by default.

Keep the output of the changelog that is currently displayed in 'mr
create' which more useful than the author-centric format from
git-format-patch.  The commit age information, however, can be dropped.

Add --cover-letter option to the 'mr create' command.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>