mr note: Provide more information in comment message
lab commit bd1c969eed38 ("mr: Migrate commands to use MR mapping code")
introduced MR context commands to lab, removing the need to enter in an
Merge Request ID if the command was being executed on a tracked branch.
After this code was introduced it is easy for a user to inadvertently
write to the wrong MR using, for example, 'lab mr note'.

It is simple to provide additional context in the editor, for example,

	"This comment is being applied to OPEN Merge Request 52"

to avoid any confusion.

Update the default note editor message to output the status and MR number
being commented on.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>