version: remove PersistentPreRun
The commit 9bf24520ef53 ("lab: Add LabPersistentPreRun() to all commands")
added the PreRun callback to the version command, which doesn't have any
flags attached to it, thus the PersistentPreRun should also not be called
for that.  By default, the --help flag is added to all commands added to the
RootCmd, therefore cobra considers the existence of "lab version --help",
which isn't true, breaking the LabPersistentPreRun call.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>