show: Add system note output
The information in the system notes is important.  For example, changing
a merge request description will generate a system note.  Engineers find
this useful in the WebUI and it is easy to show in 'lab mr show' or
'lab issue show'.

System notes cannot have replies so there is no reason to output the
note ID.  Instead, preface system notes with an asterisk '*'.

This change adds lines like, for example,

     * prarit changed the description at 2020-11-12 20:36:39.425 +0000 UTC

to the lab show commands.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>