util: Support getting MR ID from branch name
If no MR ID is specified on the command line, the mr subcommands fall
back to looking for a MR associated with the current branch. Extend
that behavior to also try a branch name specified on the command line,
as branch names are usually easier to remember than MR numbers.