show: check output type before rendering Markdown
Make sure the stdout and stderr file descriptors are refering to a real
terminal and not a pipe. If it's an usual terminal, it's safe rendering the
markdown content of an issue or mr, however, in a pipe scenario it's
possible the destination doesn't understands the rendering results (ANSI
chars).

The current tests for both `issue show` and `mr show` are already pipes, so
their logic weren't changed, however, the resulting text was, since no more
`stipansi()` is required.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>