mr_edit: Support (un)marking merge requests as draft
Gitlab supports marking a merge request as draft to indicate that it is
not ready to be merged. This is supported in the web UI (actions to mark
as draft/ready, disabled merge button in draft MRs), but not explicitly
exposed in the API: Drafts are simply indicated by prefixing the MR title
with "Draft:", "[Draft]" or "(Draft)" (or previously "WIP)[0].

lab therefore kind of supports the feature already, but it still seems
reasonable to export the feature explicitly rather than via manipulating
the title (although that's of course what the implementation does under
the hood).

[0] https://docs.gitlab.com/ee/user/project/merge_requests/work_in_progress_merge_requests.html