fork: enable different target name, namespace and path
Gitlab API allows the user to specify different destination name, namespace
(if usr has permission) and the path for the projects being forked, however,
`lab` isn't aware in the current status. This patch adds three different
flags to `lab fork` command, being that each represents a different property
in the destination location of the fork:

  -n, --name string        fork project with a different name
  -m, --namespace string   fork project in a different namespace
  -p, --path string        fork project with a different path

"name" is the new name; "namespace" is the new namespace, different from the
own user namespace; "path" is the last part of the git url, which may be
different from project's name (in case "name" isn't specified, "path" is
used as the name).

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>