fork: wait for the actual fork process completion
Today, `lab fork` doesn't check if the fork on Gitlab system is indeed
finished, it just assumes so and move on. However, some projects are big
enough to not satisfy it, for instance: the linux kernel. A project
containing the linux kernel source code may take _minutes_ to fork, and the
user will be prompted with an error if the "clone" is not skipped with
--skip-clone.

This patch forces the `lab fork` command to wait the fork completion on
Gitlab system and also adds the "--no-wait" flag if the user really don't
want to wait for anything.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>