issue_close: Allow closing as duplicate of another issue
If an issue is opened that duplicates an existing issue, it is
useful to mark it as such rather than to just close it.

Unfortunately the functionality from gitlab's web UI isn't exposed
in the API (or I'm too clumsy to find it), but we can get around
that limitation by leveraging the corresponding quick action.