ci_view: Fix screen freeze
After the tcell/v2 update, any key press during `lab ci view` results
in a lock up. It is unclear whether the changes in input handling on
the tcell side are intentional, but we can address the issue by making
the channel that notifies about input events asynchronous and ensure
that we don't overrun its queue.