util: Always override config file flags from the command line
Currently we check whether a flag's current value differs from its
default in order to skip the value from the config file. That works
in most cases, but as a result it is not possible to override a value
from the config by explicitly passing the default value:

  [mr_show]
  comments = true

  $ lab mr show --comments=false

Fix this by checking the flag's Changed property instead.