lab: Check that LAB_CORE_TOKEN is set before querying user for a token
chenrui333 noted in #558 that the goreleaser test for lab was failing.
The test is a simple execution of the git wrapper capabilities of lab.

The test does:

	git init
	touch haunted
	touch house
	git add haunted house
	git commit -a -m "Initial Commit"
	lab ls-files

and expects the output of the last command to be

	haunted
	house

The test currently fails on the command line with

[prarit@prarit dummy (master)]$ LAB_CORE_HOST=foo LAB_CORE_USER=bar LAB_CORE_TOKEN=baz lab  ls-files
Create a token with scope 'api' here: profile/personal_access_tokens
Enter default GitLab token, or leave blank to provide a command to load the token:

instead of displaying the list of files.  This occurs because the code
does not check to see if LAB_CORE_TOKEN is set before querying the user
for a token.

Check that LAB_CORE_TOKEN is set before querying the user for the token.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Reported-by: @chenrui333 (in #558)
Cc: Bruno Meneguele <bmeneg@redhat.com>