mr_list: add flags for checking merge conflicts
Allow the user to request for a list of MRs that are/aren't conflicting with
the corrent target branch. The default behavior, no flags on cli, is to present
all MRs, regardless their merge status, this patch adds two additional flags:
`--conflicts` and `--no-conflicts` are modifiers of this default behavior.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>