issue_list: Add flag for matching exact terms
Gitlab support matching on the exact search terms by enclosing the
search strings in double quotes (as is common for web search engines
as well).

This requires careful quoting or escaping on the command line - for
example '"foo bar"' or "\"foo bar\"" - which is cumbersome; make
this more convenient by adding a dedicated command line flag that
takes care of the quoting.