mr_list: Fix test
The test for milestone support creates and closes a MR on origin,
with the result that the test case for closed MRs ended up being
pushed off the list, given that only a limited number is shown
by default.

List all closed MRs to make it appear in the listing again.