mr_create: fix the target remote
The remote used as a target can be given in the command line with:

  lab mr create remote/repository/name

It was however not taken into account when comparing the branch to merge
and the target branch, and when making the cover letter template.
Instead the initial forked from branch was used, which would be invalid
if the repository wasn't forked at first from the target one. Fix this.

Signed-off-by: Antoine Tenart <atenart@kernel.org>