root: allow to configure a default remote in the configuration
When using Git repositories with many remotes, the guessDefaultRemote
logic can be wrong as it looks for the origin/upstream remote name, or
at specific branch names such as main/master. But a single repository
can have multiple remotes, and in some cases the default remote could be
another one.

This patch adds a core configuration parameter ('core.default_remote')
to allow the user defining what should be the default remote.

Using this configuration option in the global configuration file isn't
ideal. But the configuration file can be set in the .git directory of
the repositories directly. And this also works when using worktrees, a
per-worktree configuration file can be set! In the two later cases it
makes perfect sense to configure the default remote.

After this patch there is still the possibility to use custom remotes:

  # Using the default remote in the configuration if defined, or a
  # guessed one otherwise
  $ lab mr list

  # Using the one given in the cmd line
  $ lab mr list my/remote/name

Signed-off-by: Antoine Tenart <atenart@kernel.org>