go.mod: update tview module
A new update in the tview module is necessary due to an indirect update in
uniseg, which breaks the current tview being used. Even though it doesn't
really is necessary for GO111MODULE=on envs, it's nice to keep a
compatibility with GO111MODULE=off installations, like happens in some
distros (Fedora).

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>