mr_create: fix slice out of bounds for MERGE message
For enabling the draft mode in Gitlab we have specific string that must be
in the title's prefix. The code wasn't checking if the title had this
minimum size. This patch adds this check with a new helper function to avoid
a panic in case the user uses a text smaller than that.

Reported-by: Florian Muellner <fmuellner@gnome.org>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>