mr_create: Add number of commits to create metadata
First time GitLab users can make mistakes that result in a large merge
request being generated.  For example, new users may get the target branch
wrong, or push a branch that hasn't been updated, etc.  These errors always
result in way more commits than intended being pushed.

After a push to one of their hosted repositories, GitLab provides a URL
to a webpage that allows users to change the target branch and look at the
commits that are going to be pushed.  One convenient value on that page is
the number of commits that are being pushed.  In order to prevent users
from making simple mistakes, like the above, it would be good to also
provide that information the default 'lab mr create' message.

Provide the number of commits in 'lab mr create'.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>