gitlab: fix project lookup during fork
During a fork process, lab doesn't check if a fork with same name under
user's namespace is a fork from the same project the user is trying to fork
now. If the same name matches, it only return a reference for that project,
regardless where it came from.

This patch adds a check step to make sure the found project is indeed a fork
for the same project being forked. Otherwise, it returns an error informing
the user.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>