fork: allow user to set custom remote name
Today, lab only allows the user to setup the <user> or "upstream" remotes in
a fork or clone process. This patch enables the option `--remote-name` (or
`-r`) for setting a custom remote name. It's useful when the user want to
have, in the same folder, different forks of the same project (possibly
for different versions, or one for general testing and another for personal
usage, or even for using with worktrees), which is not a so unusual setup
for big projects.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>