git: skip to url in case of an empty pushurl
Some developers like to have the <pushurl> set to nothing (empty) to avoid
pushing by mistake to that remote (which for any reason he has permission to
push). In these cases, `lab` should be able to check for the default <url>
set as the pull url. This patch adds a check for empty string against
<pushurl>.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>