lab: Clean up IntVarP usage
A user noticed an issue in which 'lab mr --list' and 'lab mr list'
returned lists of 10 and 20 entries respectively.  Both effectively call
the same code in the mr_list file and the only difference is that
LabPersistentPreRun() is run on 'lab mr' and 'lab mr list' respectively.

It appears that IntVarP is not handled properly by viper (there are a few
reports of other issues with IntVarP online) and the suggestion is to
use StringVarP in place of IntVarP.

The IntVarP value is used as the number of entries returned per page
for the list commands of MRs, Issues, Projects, and Snippets.  In all of
these cases the user modified value was not being honoured in the
code so fixes for these are also included.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>