mr,issue,snippet: make shorthand flags deprecated
All commands with --list, --close and --browse flags, also has the actual
subcommand for each of them. I've seen users getting confused because the
flags in the `list` command can't be used with the shorthand, and tbh,
typing `list` is easier than `--list` (the same is valid for the others
flags).

IMHO the shorthand as a whole should be deprecated, avoiding user confusion
and, possibly, dead code for getting maintained.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>