mr show: Display output for approvers and reviewers
Several users have requested that the list of approvers and reviewers be
added to the output of the 'lab mr show' command.  This list is useful for
authors to know who to request reviews from, and to see the eligible list
of approvers.

Add Approvers, Approval Groups, and Reviewers to the output of 'lab mr
show'.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>