mr checkout: Add --force option
Users have reported a minor annoyance with 'lab mr checkout':

- reviewer checkouts out MR for review
lab mr checkout 3
- reviewer requests changes
- author makes changes
- reviewer checkouts out MR for re-review
lab mr checkout 3

and is returned a "ERROR: mr 3 branch branchname already
exists."

In most cases that I can think of the person checking out the MR
wants the latest version of the the code in the MR, not old code.  It
would be convenient to have a --force option that would delete and
overwrite the existing MR branch.

Add 'lab mr checkout --force' option that overwrites an existing MR branch
checkout.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>