mr_show: fix line wrapping
The glamour library has, by default, line wrapping activated at the mark of
80 columns, causing a differente rendering from what is seen in the WebUI.
Setting this option to `0` turns it off and prints messages as they were
exactly sent by the user, even if it was sent with `--force-linebreak`.

What is a bit confusing is that setting it to `0` make the lib to honor the
line breaks as they were sent. While setting it to a big number, e.g. 150,
won't honor the linebreaks.

This patch adds a common function for getting the terminal renderer with
line wrapping turned off and a variable style that can be passed as an
argument.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>