mr_create: Check for zero commits in a Merge Request
It is possible to open up a Merge Request with zero commits by pushing a
branch with no changes and then executing 'lab mr create'.

lab should abort the creation of an MR in this case.

Check the number of commits in the remote branch, and if it is zero, abort
the creation of the Merge Request.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>