mr create: Use remote source branch for commit logs
As reported in issue 635, in some circumstances the 'mr create' command
does not show the correct information when calculating the number of
commits or generating the commit log.

This happens because the commit comparison is done using the local
source branch which may have different content than the remote source
branch.

Use the remote source branch to calculate the number of commits and the
commit log changelog.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>