mr create: Add --source option
The GitLab WebUI provides the option to set the source repository and
branch during the creation of a merge request.  This functionality is not
present in lab.

Add a --source option that takes a remote:branch as an argument.  This
option allows users to specify the source remote and branches.

Additional fixup: Fix a missing '\n'.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>