mr_test: Generalize prepare and close test functions
This issue was noticed while implementing the --source option in 'mr
create'.

The prepare function in the merge request test code has an error that is
commented out.  This is because the 'mr -d' option is no longer valid and
always throws an error.  The prepare function attempts to clean up any
existing test merge requests prior to creating new ones.  It would be
best to clean this up, avoid duplication in the code, and generalize
the call to a single function.

In addition to that I noticed that the delete function is repetitive
throughout the test code.  The delete functions should be renamed to close,
and it also can be generalized.

Fix and generalize the prepare and close test functions.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>