mr create: Allow email addresses
On large new projects users often know each other by their email
addresses and not their GitLab usernames.  GitLab provides functionality
to lookup a UserID from an email address so it is easy to incorporate
this functionality into lab.

Allow email addresses to be used in the --reviewer and --assignee fields
of 'mr create'.

This also allows email addresses to be used in the --assign and --review
fields of 'mr edit'.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>