root: return to lab when git fails handling flags
When `lab` doesn't recognize a flag to its root command, e.g. `lab --v`,
it passes to git in the hope the user wants it to passthrough. However,
when git also fails to handle the flag, it prints a failure message and
its usage message, which is quite confusing, since the user is issuing
`lab`.

This patch checks the return from git before printing anything and in
case it's a failure, let `lab` return its usage message instead of git's.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>