lab: Add support for GitLab Testing
I have been verifying testing in preparation of moving to GitLab.

These modifications allow testing to run on both GitHub and GitLab.  They
include an addition of a LAB_CORE_TEST environment variable, and some
generalzation of the code due to different users (root on GitLab, and
travis on TravisCI).

Signed-off-by: Prarit Bhargava <prarit@redhat.com>