fork test: Delete fork project on error
The fork tests are failing with

        fork_test.go:122: Forking zaquestion/fork_test project...
            2021/04/06 00:19:49 fork.go:90: POST https://gitlab.com/api/v4/projects/5794886/fork: 409 {message: {name: [has already been taken]}, {path: [has already been taken]}}

because the fork_test_name_1 project was not properly deleted in a
previous test run.  The result is that a user must now take a manual step
of deleting the fork and/or modifying code to delete the fork.  Instead
this can be done by modifying the failure path to delete the fork.

Delete the fork project on error.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>