todo: Implmement 'mr' and 'issue' commands
The GitLab WebUI allows users to mark add Todo list items from a Merge
Request pages and Issue pages.  Unfortunately the go-gitlab library
does not implenet the adding of Issues to Todo lists [1].

Implement a "mr" command that allows users to add Merge Requests as Todo
list items.  Since go-gitlab does not support the adding of Issues to Todo
lists so put in a place holder for an "issue" command.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
[1] https://github.com/xanzy/go-gitlab/pull/1130