todo: Add --pretty output
The 'todo list' output currently only shows the Todo ID, and a link to the
Todo item.  The WebUI outputs additional information such as the state of
the Merge Requests and Issues, how the Todo entry was added to the Todo
list, etc.

Add an option, --pretty, to output the additional information.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>