config: Fix MainConfig nil pointer panic
The tests were panicking on some CI setups due to a nil pointer.  This
occurs because the CI() call is made before MainConfig is initialized
which results in MainConfig lookups hitting nil pointer panics.

Fix these panics by moving the CI() check after the initializiation of
MainConfig, and update the code comment so no one ever moves it
again.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>